👥 Add group support

+++ src/message.py
@@ -7,6 +7,7 @@ from urllib.parse import parse_qs
 
 def sendVideo(url, chatId, messageId=None, userLanguage=None):
     userLanguage  = userLanguage or dbSql.getSetting(chatId, 'language')
+    chatType = 'users' if chatId > 0 else 'groups'
     
     url = url.split('/?')[0]
     url = 'https' + url if not url.startswith('http') else url
@@ -37,14 +38,15 @@ def sendVideo(url, chatId, messageId=None, userLanguage=None):
         bot.send_chat_action(chatId, 'upload_video')
 
         if setRcVideoId:
-            sent = bot.send_video(chatId, videoId, reply_markup=resultKeyboard(userLanguage, url))
+            sent = bot.send_video(chatId, videoId, reply_markup=resultKeyboard(userLanguage, url), reply_to_message_id=messageId if chatType == 'groups' else None)
             dbSql.increaseCounter('messageRequest')
         
         else:
-            sent = bot.send_video(chatId, videoId['videoId'], reply_markup=resultKeyboard(userLanguage, url))
+            sent = bot.send_video(chatId, videoId['videoId'], reply_markup=resultKeyboard(userLanguage, url), reply_to_message_id=messageId if chatType == 'groups' else None)
             dbSql.increaseCounter('messageRequestCached')
 
-        if messageId:
+        ## Delete link in personal chat
+        if messageId and chatType == 'users':
             bot.delete_message(chatId, messageId)
 
         if setRcVideoId:
@@ -55,23 +57,32 @@ def sendVideo(url, chatId, messageId=None, userLanguage=None):
     
     #! Error
     else:
-        bot.send_message(chatId, language[video['error']][userLanguage], reply_markup=socialKeyboard(userLanguage) if video['error'] in ['exception', 'unknownError'] else None)
+        bot.send_message(chatId, language[video['error']][userLanguage], reply_markup=socialKeyboard(userLanguage) if video['error'] in ['exception', 'unknownError'] else None, reply_to_message_id=messageId if chatType == 'groups' else None)
 
 #: Text handler
 @bot.message_handler(content_types=['text'])
 def message(message):
-    userLanguage  = dbSql.getSetting(message.chat.id, 'language')
+    # Group message
+    if message.chat.id < 0:
+        if message.text.startswith(tikTokDomains):
+            sendVideo(url=message.text, chatId=message.chat.id, messageId=message.id, userLanguage='english')
 
-    if floodControl(message, userLanguage):
+        elif message.text.startswith('/start'):
+            bot.send_message(message.chat.id, language['greet']['english'].format(message.from_user.first_name), reply_markup=startKeyboard('english'))
+            dbSql.setUser(message.chat.id)
+    
+    # Personal message
+    elif floodControl(message, 'english'):
         #! Start message handler
-        if message.text == '/start':
-            bot.send_message(message.chat.id, language['greet'][userLanguage].format(message.from_user.first_name), reply_markup=startKeyboard(userLanguage))
+        if message.text.startswith('/start'):
+            bot.send_message(message.chat.id, language['greet']['english'].format(message.from_user.first_name), reply_markup=startKeyboard('english'))
+            dbSql.setUser(message.chat.id)
 
         #! Get user token
         elif message.text == '/token' or message.text == '/start getToken':
             token = dbSql.getSetting(message.chat.id, 'token', 'users')
             
-            bot.send_message(message.chat.id, language['token'][userLanguage].format(token))
+            bot.send_message(message.chat.id, language['token']['english'].format(token))
 
         #! Inline query start handler
         elif message.text == '/start inlineQuery':
@@ -79,4 +90,4 @@ def message(message):
 
         #! Link message handler
         else:
-            sendVideo(url=message.text, chatId=message.chat.id, messageId=message.id, userLanguage=userLanguage)
+            sendVideo(url=message.text, chatId=message.chat.id, messageId=message.id, userLanguage='english')
